% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhs_colour_tints.R
\name{get_nhs_colour_tints}
\alias{get_nhs_colour_tints}
\title{Get's tints of the NHS identity colours}
\usage{
get_nhs_colour_tints(tints, ..., section = NULL)
}
\arguments{
\item{tints}{a vector of numbers between 0 and 1 for the percentage colour
tint to use}

\item{...}{optional names of the colours to return}

\item{section}{optional, name of the colour sections to use}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function returns tints of the NHS identity colours
}
\examples{
get_nhs_colour_tints(seq(0.0,0.8,0.2))
get_nhs_colour_tints(seq(0.0,0.8,0.2), "Blue", "Red")
get_nhs_colour_tints(seq(0.0,0.8,0.2), section = "blues")
}
