% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhs_colours.R
\name{get_nhs_colours}
\alias{get_nhs_colours}
\title{Get's the NHS identity colours}
\usage{
get_nhs_colours(..., section = NULL)
}
\arguments{
\item{...}{optional names of the colours to return}

\item{section}{optional, name of the colour sections to use}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function gets the colours as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
get_nhs_colours()
get_nhs_colours("Blue")
get_nhs_colours("Blue", "Red")
get_nhs_colours(section = "blues")
}
