% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhs_palette.R
\name{get_nhs_palette}
\alias{get_nhs_palette}
\title{Return function to interpolate an Nhs theme colour palette}
\usage{
get_nhs_palette(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of section in Nhs theme}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an Nhs theme colour palette
}
