% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNhsColours.R
\name{getNhsColours}
\alias{getNhsColours}
\title{Get's the NHS identity colours}
\usage{
getNhsColours(..., section = NULL)
}
\arguments{
\item{...}{optional names of the colours to return}

\item{section}{optional, name of the colour sections to use}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function gets the colours as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
getNhsColours()
getNhsColours("Blue")
getNhsColours("Blue", "Red")
getNhsColours(section = "blues")
}
