% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsr_presentation.R
\name{nhsr_presentation}
\alias{nhsr_presentation}
\title{An R Markdown output format for NHS themed remark.js slides}
\usage{
nhsr_presentation(...)
}
\arguments{
\item{...}{additional arguments to pass to \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
This output format is based off of \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}} and produces an
HTML file that contains the Markdown source (knitted from R Markdown) and
JavaScript code to render slides.
}
\details{
The theme is based off of the \href{https://www.england.nhs.uk/nhsidentity/}{NHS Identity Guidelines}
produced by NHS England. It does not follow the guidelines completely, for
instance we do not use the fonts suggested by the guidelines for these are
proprietary and cannot be included in this package.
}
\examples{
\dontrun{
library(rmarkdown)

# simple invocation
render("input.Rmd", su_presentation())
}
}
