% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ.R
\docType{package}
\name{nhsrtheme}
\alias{nhsrtheme}
\alias{nhsrtheme-package}
\title{\code{nhsrtheme} package}
\description{
NHS Theme
}
\details{
#' See the README on
\url{https://github.com/tomjemmet/nhsrtheme/}
}
