% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhstheme_pal.R
\name{nhstheme_pal}
\alias{nhstheme_pal}
\title{Return function to interpolate a drsimonj color palette}
\usage{
nhstheme_pal(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a drsimonj color palette
}
