% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plot.R
\name{sample_plot}
\alias{sample_plot}
\title{Show's the NHS identity colours}
\usage{
sample_plot()
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function shows the available colours in the NHS theme, as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
sample_plot()
}
