% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_nhs.R
\name{scale_colour_nhs}
\alias{scale_colour_nhs}
\title{Color scale constructor for NHS theme colours}
\usage{
scale_colour_nhs(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in NHS theme}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_colour_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for NHS theme colours
}
