% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nhs_theme_colour_tints.R
\name{show_nhs_theme_colour_tints}
\alias{show_nhs_theme_colour_tints}
\title{Show's the NHS identity colours}
\usage{
show_nhs_theme_colour_tints(tints, colour)
}
\arguments{
\item{tints}{a vector of numbers between 0 and 1 that represents the
percentage tints required}

\item{colour}{the name of the colour to show tints of}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function shows the tints of a colour from the Nhs Theme
}
\examples{
show_nhs_theme_colour_tints(seq(0,1,0.1), "Blue")
}
