% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nhs_theme_colours.R
\name{show_nhs_theme_colours}
\alias{show_nhs_theme_colours}
\title{Show's the NHS identity colours}
\usage{
show_nhs_theme_colours()
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function shows the available colours in the NHS theme, as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
show_nhs_theme_colours()
}
