% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showNhsThemeColours.R
\name{showNhsThemeColours}
\alias{showNhsThemeColours}
\title{Show's the NHS identity colours}
\usage{
showNhsThemeColours()
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function shows the available colours in the NHS theme, as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
showNhsThemeColours()
}
