% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showNhsThemeColourTints.R
\name{showNhsThemeColourTints}
\alias{showNhsThemeColourTints}
\title{Show's the NHS identity colours}
\usage{
showNhsThemeColourTints(tints, colour)
}
\arguments{
\item{tints}{a vector of numbers between 0 and 1 that represents the
percentage tints required}

\item{colour}{the name of the colour to show tints of}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function shows the tints of a colour from the Nhs Theme
}
\examples{
showNhsThemeColourTints(seq(0,1,0.1), "Blue")
}
