% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhs.R
\name{theme_nhs}
\alias{theme_nhs}
\title{Add NHS theme to ggplot chart}
\usage{
theme_nhs()
}
\description{
This adds a theme to your chart to follow the
\href{https://digital.nhs.uk/about-nhs-digital/corporate-information-and-documents/nhs-digital-style-guidelines/how-we-look/charts}{NHS Identity}
}
