% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{getNhsColours}
\alias{getNhsColours}
\title{Get's the NHS identity colours}
\usage{
getNhsColours(section)
}
\arguments{
\item{section}{a character, or list of characters of the colour sections to
use}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function gets the colours as defined by the
\href{https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}{NHS Identity}
}
\examples{
getNhsColours()
getNhsColours("blues")
getNhsColours(c("blues", "neutrals"))
}
