% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{getNhsColourTints}
\alias{getNhsColourTints}
\title{Get's tints of the NHS identity colours}
\usage{
getNhsColourTints(tints, section = NULL)
}
\arguments{
\item{tints}{a vector of numbers between 0 and 1 for the percentage colour
tint to use}

\item{section}{a character, or list of characters of the colour sections to
use}
}
\value{
named vector of hexadecimal strings of colours
}
\description{
This function returns tints of the NHS identity colours
}
\examples{
getNhsColourTints(seq(0.0,0.8,0.2))
getNhsColourTints(seq(0.0,0.8,0.2), "blues")
}
